requirements_fine_tuning.txt
# Azure Fine-tuning Pipeline Requirements
# Dependencies for Azure OpenAI fine-tuning and OpenSource model demo

# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Azure Services
azure-identity>=1.12.0
azure-ai-ml>=1.8.0
azure-storage-blob>=12.17.0
azure-cognitiveservices-speech>=1.31.0
openai>=1.3.0

# Image Processing (for GIF export)
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# HTTP and API
requests>=2.31.0
httpx>=0.24.0

# Data Processing
pyarrow>=12.0.0
polars>=0.19.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU Support
# cuda-python>=12.0.0  # Uncomment if using CUDA
# cupy-cuda12x>=12.0.0  # Uncomment if using CUDA 12.x
